+++ backend/Portal.Persistence-EF-Core/Repositories/BookRepository.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Portal.Application.Interfaces.OuterImpl;
 using Portal.Domain.Entities;
+using Portal.Domain.Exceptions;
 using Portal.Persistence_EF_Core.Repositories.Abstract;
 using Portal.Persitence_EF_Core.FrameworkEntities;
 using System;
@@ -31,10 +32,10 @@ namespace Portal.Persistence_EF_Core.Repositories
             var bookEntity = await _context.Materials.OfType<Book>()
                 .FirstOrDefaultAsync(u => u.Id == entity.Id);
 
-            //if (userEntity == null)
-            //{
-            //    throw new NotFoundException(nameof(User), userId);
-            //}
+            if (bookEntity == null)
+            {
+                throw new EntityNotFoundException(nameof(Book));
+            }
 
             _context.Materials.Remove(bookEntity);
             return bookEntity.Id;
@@ -42,8 +43,7 @@ namespace Portal.Persistence_EF_Core.Repositories
 
         public async Task<ICollection<BookDomain>> Read()
         {
-            var books = await _context.Materials.OfType<Book>()
-                .ToListAsync();
+            var books = _context.Materials.OfType<Book>().ToList();
 
             return books.Select(x => _mapper.Map<BookDomain>(x)).ToList();
         }
@@ -55,17 +55,21 @@ namespace Portal.Persistence_EF_Core.Repositories
 
         public async Task<BookDomain> Update(BookDomain entity)
         {
-            var bookEntity = await _context.Materials.OfType<Book>()
-                .FirstOrDefaultAsync(u => u.Id == entity.Id);
+            var bookEntity = _context.Materials.OfType<Book>()
+                .FirstOrDefault(u => u.Id == entity.Id);
 
             var data = _mapper.Map<Book>(entity);
+            
+            if (bookEntity == null)
+            {
+                throw new EntityNotFoundException(nameof(Book));
+            }
 
             bookEntity.Title = data.Title;
             bookEntity.Authors = data.Authors;
             bookEntity.PageCount = data.PageCount;
             bookEntity.Format = data.Format;
             bookEntity.Published = data.Published;
-            bookEntity.Perks = data.Perks;
             bookEntity.UpdatedAt = data.UpdatedAt;
 
             _context.Materials.Update(bookEntity);
